+++ models/peprmt/R/PEcAn.PEPRMT-package.R
@@ -1,0 +1,7 @@
+#' PEcAn.PEPRMT-package
+#'
+#' PEcAn support for the tidal biochemistry model PEPRMT
+#'
+#' @importFrom dplyr all_of select %>% .data
+#' @keywords internal
+"_PACKAGE"